+++ DEV/scripts/ppr_weights.R
@@ -44,6 +44,6 @@
   weight.obic <- rbind(weight.obic, OPI_GW, fill = TRUE, use.names = TRUE)
 
   # set groundwater rechare weight to 1 for all landuse
-  weight.obic[variable %in% c('I_H_GWR', 'D_OPI_GW'), c('weight_nonpeat', 'weight_peat') := 1]
+  weight.obic[variable %in% c('I_E_GWR', 'D_OPI_GW'), c('weight_nonpeat', 'weight_peat') := 1]
     
 usethis::use_data(weight.obic, version = 3, overwrite = TRUE, compress = 'xz')